MINIFICPP-215: Make libCuRL containing classes extensions.
MINIFICPP-215: Add conditional to HTTP-CURL

MINIFICPP-215: Update Readme

MINIFICPP-215: Udate tests and ensure extensions doc is up to date

MINIFICPP-215: Add header to Extensions

MINIFICPP-215: Update linux breakage
