+++ CMakeLists.txt
@@ -53,6 +53,7 @@ set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -fsanitize=address")
 
 # Search for threads
 find_package(Threads REQUIRED)
+find_package(CURL)
 
 # Set the right openssl root path
 if (${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
@@ -74,7 +75,6 @@ list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 
 # Include LevelDB
 find_package (Leveldb REQUIRED)
-find_package(CURL REQUIRED)
 if (LEVELDB_FOUND)
     include_directories(${LEVELDB_INCLUDE_DIRS})
 else ()
@@ -109,6 +109,10 @@ add_subdirectory(thirdparty/yaml-cpp-yaml-cpp-0.5.3)
 add_subdirectory(thirdparty/civetweb-1.9.1 EXCLUDE_FROM_ALL)
 include_directories(thirdparty/concurrentqueue)
 add_subdirectory(libminifi)
+if (CURL_FOUND AND NOT DISABLE_CURL)
+	add_subdirectory(extensions/http-curl)
+	mark_as_advanced(HTTP-CURL)
+endif(CURL_FOUND)
 add_subdirectory(main)
