+++ libminifi/include/core/FlowConfiguration.h
@@ -30,7 +30,6 @@
 #include "processors/TailFile.h"
 #include "processors/ListenSyslog.h"
 #include "processors/GenerateFlowFile.h"
-#include "processors/InvokeHTTP.h"
 #include "processors/ListenHTTP.h"
 #include "processors/LogAttribute.h"
 #include "processors/ExecuteProcess.h"
@@ -70,6 +69,9 @@ class FlowConfiguration : public CoreComponent {
         logger_(logging::LoggerFactory<FlowConfiguration>::getLogger()) {
     controller_services_ = std::make_shared<core::controller::ControllerServiceMap>();
     service_provider_ = std::make_shared<core::controller::StandardControllerServiceProvider>(controller_services_, nullptr, configuration);
+    for(auto sl_func : statics_sl_funcs_){
+      registerResource("",sl_func);
+    }
   }
 
   virtual ~FlowConfiguration();
@@ -118,7 +120,20 @@ class FlowConfiguration : public CoreComponent {
     return service_provider_;
   }
 
+  static bool add_static_func(std::string functor){
+    statics_sl_funcs_.push_back(functor);
+    return true;
+  }
+
  protected:
+
+  void registerResource(const std::string &resource_function) {
+    core::ClassLoader::getDefaultClassLoader().registerResource("", resource_function);
+  }
+
+  void registerResource(const std::string &resource_location, const std::string &resource_function) {
+    core::ClassLoader::getDefaultClassLoader().registerResource(resource_location, resource_function);
+  }
 
   // service provider reference.
   std::shared_ptr<core::controller::StandardControllerServiceProvider> service_provider_;
@@ -136,6 +151,7 @@ class FlowConfiguration : public CoreComponent {
 
  private:
   std::shared_ptr<logging::Logger> logger_;
+  static std::vector<std::string> statics_sl_funcs_;
 };
 
 } /* namespace core */
