+++ main/CMakeLists.txt
@@ -45,7 +45,21 @@ include_directories(${OPENSSL_INCLUDE_DIR})
 
 # Link against minifi, yaml-cpp, civetweb-cpp, uuid, openssl, jsoncpp and leveldb
 target_link_libraries(minifiexe minifi yaml-cpp c-library civetweb-cpp ${JSON_CPP_LIB} ${UUID_LIBRARIES} ${LEVELDB_LIBRARIES} ${OPENSSL_LIBRARIES})
+message("**** HTTP Curl is from libminifi ${HTTP-CURL}")
+
+if (APPLE)
+	if (HTTP-CURL)    
+    	target_link_libraries (minifiexe -Wl,-all_load ${HTTP-CURL})
+    endif()
+else ()
+	if (HTTP-CURL)
+	  target_link_libraries (minifiexe -Wl,--whole-archive ${HTTP-CURL} -Wl,--no-whole-archive)
+	endif()    
+endif ()
+
 set_property(TARGET minifiexe PROPERTY INTERPROCEDURAL_OPTIMIZATION True)
+#SET (MYLIB -Wl,-all_load minifi-accumulo)
+#TARGET_LINK_LIBRARIES(minifiexe ${MYLIB} )
 set_target_properties(minifiexe
         PROPERTIES OUTPUT_NAME minifi)
